requirements.txt
openai>=1.14.2
requests
fastapi
uvicorn
python-dotenv
termcolor